+++ chocolatey-visualstudio.extension/extensions/Generate-UninstallArgumentsString.ps1
@@ -5,9 +5,10 @@
         [Parameter(Mandatory = $true)]
         [hashtable] $parameters,
         [string] $logFilePath,
-        [switch] $assumeNewVS2017Installer
+        [switch] $assumeNewVS2017Installer,
+        [bool] $supportsNoRestart
     )
-    Write-Debug "Running 'Generate-UninstallArgumentsString' with logFilePath:'$logFilePath', assumeNewVS2017Installer:'$assumeNewVS2017Installer'";
+    Write-Debug "Running 'Generate-UninstallArgumentsString' with logFilePath:'$logFilePath', assumeNewVS2017Installer:'$assumeNewVS2017Installer', supportsNoRestart:'$supportsNoRestart'";
     if ($assumeNewVS2017Installer)
     {
         if ($logFilePath -ne '')
@@ -16,7 +17,11 @@
         }
 
         $argumentSet = $parameters.Clone()
-        $argumentSet['norestart'] = ''
+        if ($supportsNoRestart)
+        {
+            $argumentSet['norestart'] = ''
+        }
+
         if (-not $argumentSet.ContainsKey('quiet') -and -not $argumentSet.ContainsKey('passive'))
         {
             $argumentSet['quiet'] = ''